add general css modifiers
